Add sql files and new entities

+++ src/main/java/com/northpole/snow/todo/domain/PasazerRepository.java
@@ -2,4 +2,5 @@ package com.northpole.snow.todo.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PasazerRepository extends JpaRepository<Pasazer, Long> {}
+public interface PasazerRepository extends JpaRepository<Pasazer, Integer> {
+  }
