+++ src/main/java/com/northpole/snow/todo/domain/RelacjaRepository.java
@@ -2,4 +2,5 @@ package com.northpole.snow.todo.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RelacjaRepository extends JpaRepository<Relacja, Long> {}
+public interface RelacjaRepository extends JpaRepository<Relacja, Integer> {
+}
